Enhanced the characters list by having the status collection as a header to the tableview
