+++ RickAndMortyCharacters/Presentation/Views/List/CharactersListViewController.swift
@@ -10,18 +10,15 @@ import UIKit
 class CharactersListViewController: UIViewController {
 
     // MARK: - UI Outlets
-    @IBOutlet weak var statusCollectionView: UICollectionView!
     @IBOutlet weak var charactersTableView: UITableView!
 
     // MARK: - Properties
     private let charactersTableViewCellIdentifier = "CharacterTableViewCell"
-    private let statusCollectionViewCellIdentifier = "CharacterStatusCollectionViewCell"
 
     override func viewDidLoad() {
         super.viewDidLoad()
         self.title = "Characters"
         setupTableView()
-        setupCollectionView()
     }
 
     // MARK: - UI Setup methods
@@ -29,11 +26,6 @@ class CharactersListViewController: UIViewController {
         charactersTableView.delegate = self
         charactersTableView.dataSource = self
         charactersTableView.registerCell(charactersTableViewCellIdentifier)
-    }
-    private func setupCollectionView() {
-        statusCollectionView.delegate = self
-        statusCollectionView.dataSource = self
-        statusCollectionView.registerCell(statusCollectionViewCellIdentifier)
     }
 }
 // MARK: - Tableview delegate methods
@@ -49,20 +41,8 @@ extension CharactersListViewController: UITableViewDelegate, UITableViewDataSour
         }
         return cell
     }
-
-}
-// MARK: - Collectionview delegate methods
-extension CharactersListViewController: UICollectionViewDataSource, UICollectionViewDelegate {
-
-    func collectionView(_ collectionView: UICollectionView, numberOfItemsInSection section: Int) -> Int {
-        4
+    func tableView(_ tableView: UITableView, viewForHeaderInSection section: Int) -> UIView? {
+        let headerView: StatusHeaderView = StatusHeaderView.fromNib()
+        return headerView
     }
-    
-    func collectionView(_ collectionView: UICollectionView, cellForItemAt indexPath: IndexPath) -> UICollectionViewCell {
-        guard let cell = collectionView.dequeueReusableCell(withReuseIdentifier: statusCollectionViewCellIdentifier, for: indexPath) as? CharacterStatusCollectionViewCell else {
-            return UICollectionViewCell()
-        }
-        return cell
-    }
-
 }
