+++ RickAndMortyCharacters/Presentation/Views/List/CharactersListViewController.xib
@@ -12,7 +12,6 @@
         <placeholder placeholderIdentifier="IBFilesOwner" id="-1" userLabel="File's Owner" customClass="CharactersListViewController" customModule="RickAndMortyCharacters" customModuleProvider="target">
             <connections>
                 <outlet property="charactersTableView" destination="TLa-HZ-ANb" id="fDz-3K-pMC"/>
-                <outlet property="statusCollectionView" destination="LmP-1Z-ZDg" id="T4R-aK-RIS"/>
                 <outlet property="view" destination="i5M-Pr-FkT" id="sfx-zR-JGt"/>
             </connections>
         </placeholder>
@@ -21,33 +20,18 @@
             <rect key="frame" x="0.0" y="0.0" width="393" height="852"/>
             <autoresizingMask key="autoresizingMask" widthSizable="YES" heightSizable="YES"/>
             <subviews>
-                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="plain" separatorStyle="none" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="TLa-HZ-ANb">
-                    <rect key="frame" x="0.0" y="159" width="393" height="693"/>
+                <tableView clipsSubviews="YES" contentMode="scaleToFill" alwaysBounceVertical="YES" style="grouped" separatorStyle="none" rowHeight="-1" estimatedRowHeight="-1" sectionHeaderHeight="-1" estimatedSectionHeaderHeight="-1" sectionFooterHeight="-1" estimatedSectionFooterHeight="-1" translatesAutoresizingMaskIntoConstraints="NO" id="TLa-HZ-ANb">
+                    <rect key="frame" x="0.0" y="59" width="393" height="793"/>
                     <color key="backgroundColor" systemColor="systemBackgroundColor"/>
                 </tableView>
-                <collectionView clipsSubviews="YES" multipleTouchEnabled="YES" contentMode="scaleToFill" dataMode="none" translatesAutoresizingMaskIntoConstraints="NO" id="LmP-1Z-ZDg">
-                    <rect key="frame" x="16" y="83" width="361" height="60"/>
-                    <constraints>
-                        <constraint firstAttribute="height" constant="60" id="QU9-hb-LTM"/>
-                    </constraints>
-                    <collectionViewFlowLayout key="collectionViewLayout" scrollDirection="horizontal" automaticEstimatedItemSize="YES" minimumLineSpacing="10" minimumInteritemSpacing="10" id="KcW-Q4-3CJ">
-                        <size key="itemSize" width="128" height="128"/>
-                        <size key="headerReferenceSize" width="0.0" height="0.0"/>
-                        <size key="footerReferenceSize" width="0.0" height="0.0"/>
-                        <inset key="sectionInset" minX="0.0" minY="0.0" maxX="0.0" maxY="0.0"/>
-                    </collectionViewFlowLayout>
-                </collectionView>
             </subviews>
             <viewLayoutGuide key="safeArea" id="fnl-2z-Ty3"/>
             <color key="backgroundColor" systemColor="systemBackgroundColor"/>
             <constraints>
                 <constraint firstItem="TLa-HZ-ANb" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" id="6g6-no-OPL"/>
                 <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="TLa-HZ-ANb" secondAttribute="trailing" id="alK-nF-Sdk"/>
-                <constraint firstItem="LmP-1Z-ZDg" firstAttribute="leading" secondItem="fnl-2z-Ty3" secondAttribute="leading" constant="16" id="bem-h7-Ksk"/>
+                <constraint firstItem="TLa-HZ-ANb" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" id="bl4-OK-SPz"/>
                 <constraint firstAttribute="bottom" secondItem="TLa-HZ-ANb" secondAttribute="bottom" id="dg5-HN-91B"/>
-                <constraint firstItem="TLa-HZ-ANb" firstAttribute="top" secondItem="LmP-1Z-ZDg" secondAttribute="bottom" constant="16" id="hZx-AB-gIi"/>
-                <constraint firstItem="fnl-2z-Ty3" firstAttribute="trailing" secondItem="LmP-1Z-ZDg" secondAttribute="trailing" constant="16" id="iBE-G2-XRM"/>
-                <constraint firstItem="LmP-1Z-ZDg" firstAttribute="top" secondItem="fnl-2z-Ty3" secondAttribute="top" constant="24" id="iYH-j2-Eyl"/>
             </constraints>
             <point key="canvasLocation" x="130.53435114503816" y="-11.267605633802818"/>
         </view>
