+++ RickAndMortyCharacters/Supporting Files/UIView+Extensions.swift
@@ -36,4 +36,8 @@ extension UIView {
             layer.masksToBounds = newValue > 0
         }
     }
+
+    class func fromNib<T: UIView>() -> T {
+        return Bundle.main.loadNibNamed(String(describing: T.self), owner: nil, options: nil)![0] as! T
+    }
 }
